move schema into one folder
move all schemas into one folder and create a new schema called
`Shakemap` for storing information about shakemap.
