--- src/bmkg/common/schemas/__init__.py
@@ -1,6 +1,0 @@
-"""
-This module contain common schema used by bmkg api wrapper.
-"""
-from .coordinate import Coordinate
-
-__all__ = ["Coordinate"]
