--- src/bmkg/earthquake/schemas/__init__.py
@@ -1,11 +1,0 @@
-from .earthquake import Earthquake
-from .felt_earthquake import FeltEarthquake
-from .latest_earthquake import LatestEarthquake
-from .strong_earthquake import StrongEarthquake
-
-__all__ = [
-    "Earthquake",
-    "FeltEarthquake",
-    "LatestEarthquake",
-    "StrongEarthquake",
-]
