+++ src/bmkg/schemas/__init__.py
@@ -1,8 +1,14 @@
 from .area import Area
+from .coordinate import Coordinate
 from .data import Data
+from .earthquake import Earthquake
+from .felt_earthquake import FeltEarthquake
 from .forecast import Forecast
 from .humidity import Humidity
+from .latest_earthquake import LatestEarthquake
 from .name import Name
+from .shakemap import Shakemap
+from .strong_earthquake import StrongEarthquake
 from .temperature import Temperature
 from .weather import Weather
 from .weather_forecast import WeatherForecast
@@ -11,10 +17,16 @@ from .wind_speed import WindSpeed
 
 __all__ = [
     "Area",
+    "Coordinate",
     "Data",
+    "Earthquake",
+    "FeltEarthquake",
     "Forecast",
     "Humidity",
+    "LatestEarthquake",
     "Name",
+    "Shakemap",
+    "StrongEarthquake",
     "Temperature",
     "WeatherForecast",
     "Weather",
