+++ src/bmkg/schemas/coordinate.py
