+++ src/bmkg/schemas/felt_earthquake.py
