+++ src/bmkg/schemas/forecast.py
