+++ src/bmkg/schemas/humidity.py
