+++ src/bmkg/schemas/strong_earthquake.py
