+++ src/bmkg/schemas/temperature.py
