+++ src/bmkg/schemas/weather.py
