+++ src/bmkg/schemas/weather_forecast.py
