+++ src/bmkg/schemas/wind_direction.py
