+++ src/bmkg/schemas/wind_speed.py
